leetcode/0287.Find-The-Duplicate-Number/FindTheDuplicateNumber_test.go
package leetcode

import (
	"testing"
)

func Test_findDuplicate(t *testing.T) {
	type args struct {
		nums []int
	}
	tests := []struct {
		name string
		args args
		want int
	}{
		{"Example 1", args{[]int{1, 3, 4, 2, 2}}, 2},
		{"Example 2", args{[]int{3, 1, 3, 4, 2}}, 3},
		{"Example 3", args{[]int{2, 2, 2, 2, 2}}, 2},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			if got := findDuplicate(tt.args.nums); got != tt.want {
				t.Errorf("findDuplicate() = %v, want %v", got, tt.want)
			}
		})
		t.Run(tt.name, func(t *testing.T) {
			if got := findDuplicateV2(tt.args.nums); got != tt.want {
				t.Errorf("findDuplicate() = %v, want %v", got, tt.want)
			}
		})
	}
}
